+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - fixed yes/no question answering with custom prompt
 - added support for /run command
 - added support for /reasoning-effort, /think-tokens commands and showing the values in the project bar
+- added Thinking and Answer message blocks when using reasoning models
 
 ## [0.4.0]